Prevent migrations from trying to rebuild tables
- This is bad. When the engine mounts, it forces
the migrations to run without any checking to see
whether the tables already exist, making it 
impossible to reset the database

- This gem should offer a rake task to copy the
migrations into the main app.

+++ db/migrate/20140724111844_create_nifty_attachments_table.rb
@@ -1,16 +1,19 @@
 class CreateNiftyAttachmentsTable < ActiveRecord::Migration
-  
+
   def up
-    create_table :nifty_attachments do |t|
-      t.integer :parent_id
-      t.string  :parent_type, :token, :digest, :role, :file_name, :file_type
-      t.binary :data, :limit => 10.megabytes
-      t.timestamps
+    unless ActiveRecord::Base.connection.data_source_exists?("nifty_attachments")
+      create_table :nifty_attachments do |t|
+        t.integer :parent_id
+        t.string  :parent_type, :token, :digest, :role, :file_name, :file_type
+        t.binary :data, :limit => 10.megabytes
+        t.timestamps
+      end
     end
   end
-  
+
   def down
-    drop_table :nifty_attachments
+    if ActiveRecord::Base.connection.data_source_exists?("nifty_attachments")
+      drop_table :nifty_attachments
+    end
   end
-  
 end
